Move streamMetaStore tests to app fabric
