+++ cdap-data-fabric-tests/pom.xml
@@ -76,12 +76,6 @@
     </dependency>
     <dependency>
       <groupId>co.cask.cdap</groupId>
-      <artifactId>cdap-app-fabric</artifactId>
-      <version>${project.version}</version>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>co.cask.cdap</groupId>
       <artifactId>cdap-hbase-compat-0.96</artifactId>
       <version>${project.version}</version>
       <type>test-jar</type>
